fixed a lot of bugs introduced yesterday; fixed a bug in intensity distribution widget; intensity distribution is shown in log mode now
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@1473 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/VISUAL/DIALOGS/HistogramDialog.h
@@ -51,14 +51,14 @@ namespace OpenMS
 			~HistogramDialog();
 			
 			/// Returns the value of the left splitter
-			double getLeftSplitter();
+			float getLeftSplitter();
 			/// Returns the value of the right splitter
-			double getRightSplitter();
+			float getRightSplitter();
 			
 			/// Sets the value of the left splitter
-			void setLeftSplitter(double position);
+			void setLeftSplitter(float position);
 			/// Sets the value of the right splitter
-			void setRightSplitter(double position);
+			void setRightSplitter(float position);
 
 		protected:
 			HistogramWidget *mw_;
